convert array access into a terms query + generic function
Lucene itself doesn’t support to filter on an array element, therefore
an array access was resulting in the generic function filter fallback.
By using a terms query in conjunction with a generic function filter,
Lucene can pre-filter which results in a huge performance increase.

An excerpt of a rough local single node benchmark where 1 row out of 1mio matches:

+-------------------------------+-----------------------+---------------------------------------------+
| branch                        | runtime_stats['mean'] | statement                                   |
+-------------------------------+-----------------------+---------------------------------------------+
| master                        |          3349.532433  | select a from array_access where a[1] = 101 |
| s/optimize-array-access-query |            15.7341995 | select a from array_access where a[1] = 101 |
+-------------------------------+-----------------------+---------------------------------------------+